+++ react/typings/graphql.d.ts
@@ -1,6 +1,0 @@
-declare module '*.gql' {
-  import type { DocumentNode } from 'graphql'
-
-  const query: DocumentNode
-  export default query
-}